🔁 refactor [create-lag]: created react components
