+++ src/app/create_lag/page.tsx
@@ -6,19 +6,14 @@ import React, {
   useEffect,
   useState
 } from "react";
-import {
-  BsPlusLg as PlusIcon,
-  BsTelegram as TelegramIcon,
-} from "react-icons/bs";
 
 // thecoreloop Components
-import Article from "@/components/create_lag/Article";
-import CategoryHeading from "@/components/create_lag/CategoryHeading";
+import LAG_Create from "@/components/create_lag/LAG_Create";
+import LAG_Preview from "@/components/create_lag/LAG_Preview";
 
 // utils
-import { ILAG, ICategoryGroup } from "@/utils/types";
+import { ILAG } from "@/utils/types";
 import LAG_CATEGORIES from "@/utils/lag-categories";
-import { formatDate, getTodaysDate } from "@/utils/date";
 
 // LAG Context
 const LookAtGaming = createContext({});
@@ -39,148 +34,6 @@ export default function Create_LAG(): React.ReactElement {
   useEffect(() => {
     console.log("LAG: ", LAG);
   }, [LAG]);
-
-  function renderCategories(): React.ReactElement[] {
-    let result: React.ReactElement[] = [];
-
-    for (let i = 0; i < LAG.content.length; i++) {
-      const category_group: ICategoryGroup = LAG.content[i];
-
-      const Articles: React.ReactElement[] = [];
-      for (let j = 0; j < category_group.articles.length; j++) {
-        // const article: IArticle = category_group.articles[j];
-        Articles.push(
-          <Article
-            key={j}
-            category={category_group.category}
-            index={j}
-            context={LookAtGaming}
-            updateLAG={updateLAG}
-          />
-        );
-      }
-
-      result.push(
-        <div key={i} className="category-container">
-          <CategoryHeading
-            key={i}
-            category={category_group.category}
-          />
-          <div className="article-list">
-            {Articles}
-          </div>
-          <button
-            key={"a" + i}
-            className="btn add-article"
-            onClick={() => addArticle(category_group.category)}
-          >
-            <PlusIcon color="white" size="24px" />
-          </button>
-        </div>
-      );
-    }
-
-    return result;
-  }
-
-  function updateLAG(LAG_param: string): void {
-    switch (LAG_param) {
-      case "num": {
-        const lag_num: HTMLInputElement = document.getElementById(
-          "lag-number"
-        ) as HTMLInputElement;
-
-        try {
-          const num: number = Number(lag_num.value);
-          setLAG({ ...LAG, num });
-          return;
-        } catch (err) {
-          console.log(err);
-        }
-      } case "date": {
-        const lag_date: HTMLInputElement = document.getElementById(
-          "lag-date"
-        ) as HTMLInputElement;
-
-        try {
-          const date: string = formatDate(lag_date.value);
-          setLAG({ ...LAG, date });
-          return;
-        } catch (err) {
-          console.log(err);
-        }
-      } case "subheading": {
-        const lag_subheading: HTMLInputElement = document.getElementById(
-          "lag-subheading"
-        ) as HTMLInputElement;
-
-        try {
-          const subheading: string = lag_subheading.value;
-          setLAG({ ...LAG, subheading });
-          return;
-        } catch (err) {
-          console.log(err);
-        }
-      } case "content": {
-        const content: ICategoryGroup[] = [];
-
-        const cgc: HTMLDivElement = document.getElementById(
-          "category-group-container"
-        ) as HTMLDivElement;
-
-        for (let i = 0; i < cgc.children.length; i++) {
-          const cc: HTMLDivElement = cgc.children[i] as HTMLDivElement;
-          const category: string = cc
-            .children[0]
-            .children[0]
-            .textContent!;
-          const category_group: ICategoryGroup = {
-            category,
-            articles: [],
-          };
-          for (let j = 0; j < cc.children[1].children.length; j++) {
-            const article = cc
-              .children[1]
-              .children[j]
-              .children[0]
-              .children;
-            category_group.articles.push({
-              caption: (
-                article[0].lastChild as HTMLTextAreaElement
-              ).value || "",
-              link: (
-                article[1] as HTMLInputElement
-              ).value || "",
-              alt_text: (
-                article[2] as HTMLInputElement
-              ).value || "",
-            });
-          }
-          content.push(category_group);
-        }
-
-        setLAG({ ...LAG, content });
-        return;
-      }
-    }
-  }
-
-  function handleSubmit(): void {
-    console.log("LAG: ", LAG);
-  }
-
-  function addArticle(category: string): void {
-    const content: ICategoryGroup[] = LAG.content;
-    const category_group: ICategoryGroup = content.filter(category_group => (
-      category_group.category == category
-    ))[0];
-    category_group.articles.push({
-      caption: "",
-      link: "",
-      alt_text: "",
-    });
-    setLAG({ ...LAG, content });
-  }
 
   return (
     <LookAtGaming.Provider value={{LAG, setLAG}}>
@@ -221,110 +74,9 @@ export default function Create_LAG(): React.ReactElement {
           </div>
 
           <div className="row-container">
-            {/* Create LAG Container */}
-            <div className="tcl-container gray">
-              <h2 className="lag-heading">
-                Create
-              </h2>
+            <LAG_Create context={LookAtGaming} />
 
-              <div className="num-date-container">
-                <div className="input-container small row">
-                  <label
-                    className="row"
-                    htmlFor="lag-number"
-                  >
-                    Look at Gaming #:
-                  </label>
-                  <input
-                    id="lag-number"
-                    className="input lag-number"
-                    type="text"
-                    maxLength={4}
-                    placeholder="1234"
-                    autoFocus={true}
-                    autoComplete="off"
-                    onChange={(event) => {
-                      event.target.value = event.target.value.trim();
-                      updateLAG("num");
-                    }}
-                  />
-                </div>
-
-                <div className="input-container small row">
-                  <label
-                    className="row"
-                    htmlFor="lag-date"
-                  >
-                    Date:
-                  </label>
-                  <input
-                    id="lag-date"
-                    className="input lag-date"
-                    type="date"
-                    onChange={() => updateLAG("date")}
-                  />
-                </div>
-              </div>
-
-              <div className="input-container column">
-                <label
-                  className="column"
-                  htmlFor="lag-subheading"
-                >
-                  Subheading:
-                </label>
-                <textarea
-                  id="lag-subheading"
-                  className="input lag-textarea"
-                  placeholder="gm gm"
-                  onChange={() => updateLAG("subheading")}
-                />
-              </div>
-
-              <div id="category-group-container">
-                {renderCategories()}
-              </div>
-
-              <button
-                className="btn lag-submit"
-                onClick={handleSubmit}
-              >
-                Submit
-              </button>
-            </div>
-
-            {/* Preview LAG Container */}
-            <div className="tcl-container gray">
-              <div className="preview-heading-container">
-                <TelegramIcon size="2rem" color="#229ED9" />
-                <h2 className="lag-heading preview">
-                  Telegram Preview
-                </h2>
-              </div>
-
-              <div className="lag-preview">
-                <h3 className="heading">
-                  {`Look at Gaming #`}
-                  <span>
-                    {LAG.num
-                      ? LAG.num
-                      : "XYZ"
-                    }
-                  </span>
-                  {` | `}
-                  <span>
-                    {LAG.date
-                      ? LAG.date
-                      : getTodaysDate()
-                    }
-                  </span>
-                </h3>
-
-                <p className="subheading">
-                  {LAG.subheading}
-                </p>
-              </div>
-            </div>
+            <LAG_Preview context={LookAtGaming} />
           </div>
         </div>
       </main>
